Add elvis config & cleanup
Adds an elvis config and some cleanup of the suite to not vomit
all over the place and hide issues that might be worth discussing.

The linter does not pass on this point, this is on purpose as we
should discuss if what rules we want to impose plus 2 errors are
based on missing callbacks.